[Backend] Changed packages. [FrontEnd] Added client file
