+++ feedback-back/pom.xml
@@ -102,7 +102,7 @@
                         <!-- add Main-Class to manifest file -->
                         <transformer
                                 implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                            <mainClass>com.feedback.back.app.Main</mainClass>
+                            <mainClass>com.feedback.app.Main</mainClass>
                         </transformer>
                     </transformers>
