--- feedback-back/src/main/java/com/feedback/back/entities/FieldType.java
@@ -1,9 +1,0 @@
-package com.feedback.back.entities;
-
-/**
- * Created by Samarth Bhargav on 16/7/15.
- */
-public enum FieldType
-{
-    NUMERICAL, CATEGORICAL, TEXT
-}
