--- feedback-back/src/main/java/com/feedback/back/entities/api/ErrorEntity.java
@@ -1,57 +1,0 @@
-package com.feedback.back.entities.api;
-
-/**
- * Created by Samarth Bhargav on 8/7/15.
- */
-public class ErrorEntity
-{
-    private String message;
-    private int statusCode;
-
-
-    public ErrorEntity()
-    {
-
-    }
-
-
-    public ErrorEntity( String message, int statusCode )
-    {
-        this.message = message;
-        this.statusCode = statusCode;
-    }
-
-
-    public String getMessage()
-    {
-        return message;
-    }
-
-
-    public void setMessage( String message )
-    {
-        this.message = message;
-    }
-
-
-    public int getStatusCode()
-    {
-        return statusCode;
-    }
-
-
-    public void setStatusCode( int statusCode )
-    {
-        this.statusCode = statusCode;
-    }
-
-
-    @Override
-    public String toString()
-    {
-        return "ErrorEntity{" +
-            "message='" + message + '\'' +
-            ", statusCode=" + statusCode +
-            '}';
-    }
-}
