+++ feedback-back/src/main/java/com/feedback/entities/FieldType.java
@@ -1,0 +1,9 @@
+package com.feedback.entities;
+
+/**
+ * Created by Samarth Bhargav on 16/7/15.
+ */
+public enum FieldType
+{
+    NUMERICAL, CATEGORICAL, TEXT
+}
