+++ feedback-back/src/main/resources/log4j.properties
@@ -1,4 +1,4 @@
-log4j.rootLogger=DEBUG,stdout
+log4j.rootLogger=INFO,stdout
 
 # Direct log messages to stdout
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
