+++ feedback-front/templates/layout.html
@@ -10,6 +10,7 @@
         <script src="{{ url_for('static', filename='js/jquery-2.1.4.min.js') }}"></script>
         <script src="{{ url_for('static', filename='js/uikit.min.js') }}"></script>
         <script src="{{ url_for('static', filename='js/utils.js') }}"></script>
+        <script src="{{ url_for('static', filename='js/feedback.js') }}"></script>
     </head>
 
     <body>
